change refresh time for new booking and driver steps

+++ backend/src/main/java/com/joffrey/bolber/business/domain/driver/Coordinates.java
@@ -1,6 +1,6 @@
 package com.joffrey.bolber.business.domain.driver;
 
-public record Coordinates(int horizontal, int vertical) {
+public record Coordinates(float horizontal, float vertical) {
 
 
 }
